backend/requirements.txt
Django>=5.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.3.1
django-filter>=25.1 